cmd/internal/obj/sparc64, runtime: implement new calling sequence
for leaves

Make liblink use the caller's frame for NOFRAME leaves. The user
is required to specify NOFRAME. There's no automatic promotion of
framesize-0 leaves to NOFRAME.

The old framesize-$8 syntax is not accepted any more. As a sanity
check, liblink will refuse to assemble if NOFRAME is specified with
a non-zero frame size.

Update all assembly to use the new NOFRAME syntax.

Also make all trampolines NOFRAME. With the SPARC64 calling convention,
trampolines must be NOFRAME.

This change also implements runtime∕internal∕atomic·{And,Or}8 in
Go instead of assembly.

Change-Id: Iba69948bb476031726de75968f363ca4f04a3e1f