cmd/internal/obj/sparc64: make the prolog and epilog more ABI compliant
Removes the silly Sys V emulation done in the prolog. Start the
prolog with a RNOP, this keep DTrace happy.

Also make stack frames 16-byte aligned.

Use R1 in the epilog instead of TMP. We don't know if this does
anything useful, we just go the safer route of matching what gcc
does.

Change-Id: I26ba8ea86e534c80eff402cbcef54033c4b190f5