runtime: load the frame pointer saved in gobuf in recovery
Otherwise we continue execution with the wrong frame buffer. This
doesn't matter for function return, since the frame pointer is
loaded from the stack before returning, but it matters for accessing
local variables and function arguments.

We also make recovery return at PC+8, since on SPARC64 the program
counter holds the address of the current instruction (in this case
the CALL to deferreturn), but we need to fake a return from
deferreturn. This change has probably no effect, since extraneous
deferreturn calls should not cause any problems.