runtime: add regflush stub for use from Go code
shrinkstack() fails right now; unclear why but some pointers are not being
adjusted.

A register flush is likely going to be required somewhere, but I'm not sure
where to place it yet.  This adds a regflush() stub to make it easy to
invoke behind the appropriate guard when that is determined.