runtime: save caller context in runtime·sigtramp
Go saves the callee context in the callee frame during callee prolog,
but the ABI requires to save the caller context on the caller frame
in the callee prolog. To avoid the overhead of saving both the
caller and callee context in every prolog, and the overhead of
another function modifier, make the c-to-go bridge save the caller
context in the caller frame before dispatching to runtime·sigtramp1.