runtime: when unwinding the stack, don't decrement the program counter
On almost every architecture out there, the program counter helds
the address of the next instruction. When we unwind the stack, we
decrement all the saved program counters we find, so we get the
stack map, source code location, and possibly spadj corresponding
to the CALL instruction.

SPARC64 is not like this. On SPARC64 the program counter helds the
address of the current instruction, which means it already helds
the address of the CALL instruction. So on SPARC64, we don't need
to decrement the program counter as we process stack frames.

This fixes stack traces, but more importantly it seems to fix the
garbage collector.