cmd, runtime: reserve extra stack space for sparc64
sparc64 requires that 2047 bytes above the current %sp be reserved for use
by the system and refers to this as the 'stack bias'.  As such, we must
guarantee that at least that many bytes above the current sp are always
available and are never used by the Go runtime.

As such, _StackSystem must include this bias, so that runtime checks for
can ensure that BSP-framesize > stackguard.