cmd, runtime: fix stacksplit logic on sparc64
Various errors were found in the stacksplit code on sparc64; runtime
registers (REG_RT*) must be used instead of temporary registers (REG_TMP*)
as the temporary registers are often used by liblink which can lead to
unexpected clobbering of values; and the sign used when adding bias / guard
to REG_RSP was not correct.

In addition, sparc64 was not applying the nosplit stack check at all due to
a missing case for R_CALLSPARC64, and after fixing that, it was discovered
that sparc64 requires ~2100 bytes for a chain of nosplit functions due to
the 176 btye minimum frame size and lack of optimizations that require
extra temporaries.  This requires a significant increase in the size of the
StackGuard.  Also, since sparc64 has a 176-byte minimum frame size,
StackSmall must be doubled at a minimum to cover the common case. And any
updates applied to runtime/stack.go must also be applied to
cmd/internal/obj/stack.go for historical reasons.

Aram also provided additional debugging functions that make failures less
mysterious by providing current stackguard information and stack lo/hi when
a failure during moretack is encountered instead of just silently aborting.