runtime: stack debugging adjustments for sparc64
Decrease StackMin to 4096 to more aggressively trigger current stack growth
failure.

Decrease StackGuard to 2048 bytes as this seems to be enough to pass the
nosplit stack check.

Add stack delta to debugging output when finding bad pointer.

Always print lr in hex in stackDebug messages.

Include stackguard in stackDebug messages.

Add extra newlines to stackDebug messages that are printed near the end of
the function to make it easier to detect different calls to newstack.