cmd: update test to reflect new size of type Prog
The sizeof(type Prog) changed with the sparc64 port due to the increase in
the size of type As to accommodate additional instructions for sparc64.
Upstream chose a different solution that would still work that didn't
require increasing the size of As, but would require backporting other
changes that reduce the size of Prog and would complicate our porting
efforts.