runtime: fix runtime.jmpdefer and limit register usage
Fix jmpdefer on sparc64 to restore the caller of deferreturn's ILR in
addition to restoring deferreturn's OLR and BSP/BFP.

Also, since runtime.jmpdefer inserts itself between the function we want to
call and the one we actually need to return to, it's safest to only use the
REG_TMP*, REG_G* or REG_L* registers (with the exception of ILR/OLR) since
any alterations of the REG_I* or REG_O* registers can clobber input or
return values.  This shouldn't matter since the defer mechanism should only
be used with Go functions, but it costs nothing to avoid any potential
issues and appears to work as expected.