runtime: fix a float conversion failure case
On sparc64, the floating point registers are effectively per-thread, not
per-process.  So we not only need to call reginit() at startup, but for
each thread created by newosproc() (via tstart_sysvicall()).

Failure to initialize REG_YTWO with the special 2.0 value used for
conversions when starting a new 'm' lead to failure of some tests if
multiple ms were used (for example, if 'TestDelims' and 'TestNumberParse'
were run at the same time via the text/template/parse tests).