runtime: make sure tls_g is used in a single file
When assembling code, we do it one file at a time, and we need to
be aware whether a symbol is a TLS symbol in order to generate the
correct relocations for accessing the symbol.

In Go assembly, declaration is part of the definition, and we can
only define a symbol once, so every usage of a TLS symbol must
happen in the file containing the declaration.