cmd: fix relocation generation
When relocations are emitted on AMD64, the value field is always zero since
it is ignored.  However, on SPARC64, the value of the field to be
relocated, is added to the r_addend member, and serves as the relocation
addend. As such, like AMD64, the value field should also always be zero for
Go's purposes.

The GNU linker ignores the value field and treats SPARC as amd64
(incorrectly) which is why this behaviour is only observed when using the
Solaris linker.