cmd/compile/internal: implement Hmul64[u] for sparc64
Implement Hmul64[u] for sparc64 as a synthetic UMULXHI instruction since
Go no longer provides a fallback path for architectures that lack one.
VIS3 offers this as a single instruction, but we don't support VIS yet.

This is a naive implementation that does almost all of the work that
would normally be needed for a full 64-bit*64-bit multiply.

This allows part of the copy.go test to be SSA-built/run.