cmd/compile/internal: fix *MULXHI register allocation
Apparently, the SSA-based backend will happily schedule an input
register to also be an output register, so if you need scratch registers
that are guaranteed safe, you must remove them from both the input and
output register lists and mark them as clobbered.

This was discovered as I would inline or not inline the u8 function in
copy.go or make other calls which would change register scheduling and
thus make seemingly unrelated changes allow the test to pass.

This fixes additional parts of the copy.go test.