cmd/compile/internal/ssa: fix CMPconst for sparc64
CMPconst is limited to a signed 13-bit immediate for comparisons, so
don't emit CMPconst if the constant is < -4097 or > 4095.

It's unclear to me whether this should be handled in the SSA layer or
automatically by obj asm output.  It seems like handling this in the SSA
layer is preferable since more optimal register allocation can occur.

This allows the 64bit.go test to be SSA-built/run.