cmd/compile/internal: fix Move/Zero for sparc64
The rules for moving/zeroing on SPARC64 were not taking alignment into
account.  sparc64 is similar to arm64 in implementation, but with the
alignment requirements of mips64 and ppc64, so a hybrid of the two is
used here.

This allows the bigmap.go test to be SSA-built/run with substantially
less code than the old compiler backend due to zeroing optimizations.