cmd/compile/internal/sparc64: fix BlockRetJmp generation
A BlockRetJmp should be transformed into an ARET, not AJMP, so that the
obj backend for sparc64 will properly detect the special case.

This fix allows test/ddd.go, and test/method.go to be SSA-built/run.

All tests that can be cross-compiled with a simple 'go tool compile' now
work as expected.