cmd/compile/internal/sparc64: fix SLLmax/SRLmax
We can't safely write to the output register(s) until we're done
referencing the input register(s) as they may be the same due to
register allocation scheduling by the SSA backend.

As such, stash the second argument in a temporary register so that
we can safely refer to it even after writing to the output register.