cmd/compile/internal/sparc64: fix float conversions
FSTOX needs the same aglorithm applied as FDTOX (use native conversion
if value less than 2^63; otherwise subtract 2^63, convert, and add
back).

FXTOS/FXTOD were not handling the unsigned case correctly, not only did
they need to copy the input value to a temporary register for
adjustments, but they also had the 2^63 comparison reversed.

FXTOS was wrongly using AFMULD to double the value after conversion for
the unsigned > 2^63 case.

This is enough for test/ken/convert.go to be SSA-built/executed.